# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BotArmy Backend",
    "summary": "Complete backend API for BotArmy multi-agent system with HITL oversight and workflow orchestration",
    "description": "BotArmy POC Backend Services - Multi-Agent Orchestration Platform",
    "contact": {
      "name": "BotArmy Development Team",
      "url": "https://github.com/your-org/bmad"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/hitl/{request_id}/respond": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Respond To Hitl Request",
        "description": "Respond to a HITL request with approve, reject, or amend actions.",
        "operationId": "respond_to_hitl_request_api_v1_hitl__request_id__respond_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlResponseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlProcessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{request_id}": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request",
        "description": "Get a specific HITL request.",
        "operationId": "get_hitl_request_api_v1_hitl__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{request_id}/history": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request History",
        "description": "Get the history of a HITL request.",
        "operationId": "get_hitl_request_history_api_v1_hitl__request_id__history_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Hitl Request History Api V1 Hitl  Request Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/project/{project_id}/requests": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Project Hitl Requests",
        "description": "Get all HITL requests for a project.",
        "operationId": "get_project_hitl_requests_api_v1_hitl_project__project_id__requests_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HitlRequestResponse"
                  },
                  "title": "Response Get Project Hitl Requests Api V1 Hitl Project  Project Id  Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/bulk/approve": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Bulk Approve Hitl Requests",
        "description": "Bulk approve multiple HITL requests.",
        "operationId": "bulk_approve_hitl_requests_api_v1_hitl_bulk_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlBulkApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlBulkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{request_id}/context": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request Context",
        "description": "Get full context for a HITL request including artifacts and task details.",
        "operationId": "get_hitl_request_context_api_v1_hitl__request_id__context_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/statistics": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Statistics",
        "description": "Get HITL statistics for monitoring and analytics.",
        "operationId": "get_hitl_statistics_api_v1_hitl_statistics_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlStatisticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/config/trigger-condition": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Configure Hitl Trigger Condition",
        "description": "Configure a HITL trigger condition.",
        "operationId": "configure_hitl_trigger_condition_api_v1_hitl_config_trigger_condition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Configure Hitl Trigger Condition Api V1 Hitl Config Trigger Condition Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HitlAction": {
        "type": "string",
        "enum": [
          "approve",
          "reject",
          "amend"
        ],
        "title": "HitlAction",
        "description": "HITL response action enumeration."
      },
      "HitlBulkApproveRequest": {
        "properties": {
          "request_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Request Ids"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "request_ids",
          "comment"
        ],
        "title": "HitlBulkApproveRequest",
        "description": "Request model for bulk HITL approval."
      },
...
      "HitlProcessResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "workflow_resumed": {
            "type": "boolean",
            "title": "Workflow Resumed"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "action",
          "status",
          "message",
          "workflow_resumed"
        ],
        "title": "HitlProcessResponse",
        "description": "Response model for HITL processing."
      },
...
      "HitlResponseRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/HitlAction"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "HitlResponseRequest",
        "description": "Request model for HITL responses."
      },
...
    }
  },
...
}